// cocreate-node/src/models/userModel.ts
export type User = {
    id: string;
    name: string;
    email: string;
    password: string;
}

export type UserProfile = {
    id: string;
    first_name: string;
    last_name: string;
    about_me: string;
    skills: string;
    job_title: string;
    experience: string;
    education: string;
    location: string;
    languages: string;
    linkedin: string;
    github: string;
}